# docs/openapi-zeonexus-control.yaml
openapi: 3.0.3
info:
  title: ZeoNexus Gateway Control API
  version: 1.1.0
  description: Internal management-network contract. All operations require ZeoNexus HMAC headers.
servers:
  - url: http://127.0.0.1:3101/internal/nexus/v1
security:
  - ZeoSignature: []
paths:
...
  /channels/{id}:
    put:
      operationId: upsertChannel
      parameters: [$ref: '#/components/parameters/ResourceId']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Channel'}
      responses:
        '200': {description: Channel accepted}
    delete:
      operationId: disableChannel
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/Revision'
      responses:
        '204': {description: Channel disabled}
...
components:
  securitySchemes:
    ZeoSignature:
      type: apiKey
      in: header
      name: X-Zeo-Signature
      description: HMAC-SHA256 with X-Zeo-Timestamp and X-Zeo-Nonce.
  parameters:
    ResourceId: {name: id, in: path, required: true, schema: {type: string, minLength: 1}}
    Revision: {name: revision, in: query, required: true, schema: {type: integer, minimum: 1}}
  schemas:
...
    Channel:
      type: object
      required: [name, type, base_url, models, status, profile, revision]
      properties:
        name: {type: string}
        type: {type: integer}
        base_url: {type: string, format: uri}
        key: {type: string, writeOnly: true}
        models: {type: array, items: {type: string}, minItems: 1}
        model_mapping: {type: object, additionalProperties: {type: string}}
        status: {type: string, enum: [active, disabled]}
        weight: {type: integer, minimum: 0}
        priority: {type: integer}
        profile: {$ref: '#/components/schemas/Profile'}
        site_id: {type: string}
        node_id: {type: string}
        endpoint_id: {type: string}
        max_concurrency: {type: integer, minimum: 0}
        revision: {type: integer, minimum: 1}
...
    Profile:
      type: string
      enum: [aggregation, inference]